egg: Fix undefined behavior on excessive left shift

+++ egg/egg-buffer.c
@@ -321,7 +321,7 @@ egg_buffer_encode_uint32 (unsigned char* buf, uint32_t val)
 uint32_t
 egg_buffer_decode_uint32 (unsigned char* ptr)
 {
-	uint32_t val = ptr[0] << 24 | ptr[1] << 16 | ptr[2] << 8 | ptr[3];
+	uint32_t val = (uint32_t) ptr[0] << 24 | ptr[1] << 16 | ptr[2] << 8 | ptr[3];
 	return val;
 }
